+++ build.R
@@ -6,7 +6,9 @@ library(forecastHybrid)
 knit("README.Rmd", "README.md")
 
 document("pkg")
+build_vignettes("pkg")
 test("pkg")
+check("pkg")
 
 build("pkg")