+++ pkg/inst/doc/forecastHybrid.R
@@ -1,0 +1,9 @@
+## ----cran_install, eval = FALSE------------------------------------------
+#  install.packages("forecastHybrid")
+
+## ----github_install, eval = FALSE----------------------------------------
+#  devtools::install_github("ellisp/forecastHybrid/pkg")
+
+## ----load_library, message = FALSE---------------------------------------
+library(forecastHybrid)
+